Added demo code to draw some pumpkins and scroll the screen (without zones at this time).

+++ object.h
@@ -9,10 +9,10 @@ public:
 	void render();
 	void setBitmapHwnd(ALLEGRO_BITMAP *bmp_handle);
 	ALLEGRO_BITMAP *bmp_handle;
-	float x;
-	float y;
-	float width;
-	float height;
+	long x;
+	long y;
+	long width;
+	long height;
 };
